модалка нашей команды
добавила модальное окно со всеми участниками и ссылками  гитхаб, анимировала сердце в футере
