+++ src/index.html
@@ -35,6 +35,7 @@
       <!-- <include src="./partials/modalRegist.html"></include> -->
     </main>
     <include src="./partials/footer.html"></include>
+    <include src="./partials/modalTeem.html"></include>
     <!-- The core Firebase JS SDK is always required and must be listed first -->
     <script src="https://www.gstatic.com/firebasejs/8.6.8/firebase-app.js"></script>
