+++ src/js/teemList.json
@@ -1,0 +1,9 @@
+[
+  { "name": "Vladymyr Rzhanov", "gitlink": "https://github.com/VladymyrRzhanov" },
+  { "name": "Andrei Lysenkovskiy", "gitlink": "https://github.com/andreilysen" },
+  { "name": "Roma Kokontsev", "gitlink": "https://github.com/RomaKokontsev" },
+  { "name": "Oleksandr Tarasyuk", "gitlink": "https://github.com/OlekSandr-offensive" },
+  { "name": "Oksana Martynenko", "gitlink": "https://github.com/oksana-mart" },
+  { "name": "Alona Nihalchuk", "gitlink": "https://github.com/AlonaNihalchuk" },
+  { "name": "Svitlana Arkhinos", "gitlink": "https://github.com/svarkhinos" }
+]
