+++ src/partials/modalTeem.html
@@ -1,0 +1,26 @@
+<div class="backdrop backdrop-hidden backdrop-teem" data-modal-teem>
+  <div class="modal modal-img">
+    <button class="close-button close-button-light" type="button" data-close-btn>
+      <svg height="30" width="30">
+        <use href="././images/symbol-defs.svg#icon-close-modal"></use>
+      </svg>
+    </button>
+    <div class="teem-container">
+      <h3 class="teem-title">Our teem</h3>
+      <ul class="teem-head-list">
+        <li class="teem-item">
+          Teem Leader
+          <a class="teem-item-link" href="https://github.com/solarenergy990" target="_blank"
+            ><span class="teem-item-style">&ensp;Eugen Soliar</span></a
+          >
+        </li>
+        <li class="teem-item">
+          Scrum Master <a class="teem-item-link" href="https://github.com/VeraBohatyrenko" target="_blank"
+          ><span class="teem-item-style">Vera Bohatyrenko</span>
+        </li>
+      </ul>
+      <p class="teem-members">Members</p>
+      <ul class="teem-members-list"></ul>
+    </div>
+  </div>
+</div>
