+++ src/sass/main.scss
@@ -16,6 +16,7 @@
 @import './components/footer';
 
 @import './components/modal';
+@import './components/modalTeem';
 // @import './components/modalRegist';
 @import './components/container';
 @import './components/theme';
