+++ src/templates/teemListTpl.hbs
@@ -1,0 +1,4 @@
+{{#each this}}
+<li><a class="teem-item-link" href="{{gitlink}}" target="_blank"
+          >{{name}}</li>
+{{/each}}
